LayoutTests/platform/ios-simulator/editing/pasteboard/drag-files-to-editable-element-expected.txt

CONSOLE MESSAGE: line 29: TypeError: undefined is not a function (evaluating 'eventSender.beginDragWithFiles(files)')
If we drag files onto an editable area, then the filenames should be inserted into the editable area.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE
